refactor : 컨트롤러 인터페이스를 통해 모든 컨트롤러를 포괄하도록 리팩토링
