+++ src/main/java/subway/controller/Controller.java
@@ -1,0 +1,5 @@
+package subway.controller;
+
+public interface Controller {
+    void run();
+}
